Refactor rpm bits into packagesystem module
Prep for cleaning this up and making it more generic
to support non-rpm systems in a pluggable way.
